requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
# Additional utilities that might be needed
requests>=2.25.0
ruff>=0.1.0
tqdm>=4.60.0

# For downloading ImageNet dataset
kagglehub
